requirements.txt
# AI Trader US - Dependencies

# Data
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=15.0.0

# Visualization
plotly>=5.18.0

# Market Calendar
exchange-calendars>=4.5.0
pytz>=2023.3

# Config & Logging
loguru>=0.7.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Async
aiohttp>=3.9.0

# Analytics
scipy>=1.11.0

# CLI
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
